geom viewer: extract drawing in separate lib/GeomDrawing.js script
